[Refact] added logger header
